# example-project/cmd/web-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  project.ProjectArchiveRequest:
    properties:
      id:
        example: 985f1746-1d9f-459f-a2d9-fc53ece5ae86
        type: string
    required:
    - id
    type: object
  project.ProjectCreateRequest:
    properties:
      account_id:
        example: c4653bf9-5978-48b7-89c5-95704aebb7e2
        type: string
      name:
        example: Rocket Launch
        type: string
      status:
        enum:
        - active
        - disabled
        example: active
        type: string
    required:
    - account_id
    - name
    type: object
  project.ProjectResponse:
    properties:
      account_id:
        example: c4653bf9-5978-48b7-89c5-95704aebb7e2
        type: string
      archived_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      created_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      id:
        example: 985f1746-1d9f-459f-a2d9-fc53ece5ae86
        type: string
      name:
        example: Rocket Launch
        type: string
      status:
        $ref: '#/definitions/web.EnumResponse'
        type: object
      updated_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
    required:
    - account_id
    - id
    - name
    type: object
  project.ProjectUpdateRequest:
    properties:
      id:
        example: 985f1746-1d9f-459f-a2d9-fc53ece5ae86
        type: string
      name:
        example: Rocket Launch to Moon
        type: string
      status:
        enum:
        - active
        - disabled
        example: disabled
        type: string
    required:
    - id
    type: object
...
  web.EnumOption:
    properties:
      selected:
        example: true
        type: boolean
      title:
        example: Active Etc
        type: string
      value:
        example: active_etc
        type: string
    type: object
  web.EnumResponse:
    properties:
      options:
        items:
          $ref: '#/definitions/web.EnumOption'
        type: array
      title:
        example: Active Etc
        type: string
      value:
        example: active_etc
        type: string
    type: object
...
  web.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/web.FieldError'
        type: array
    type: object
  web.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
  web.TimeResponse:
    properties:
      date:
        example: "2019-06-25"
        type: string
      kitchen:
        example: 3:00AM
        type: string
      local:
        example: Tue Jun 25 3:00AM
        type: string
      local_date:
        example: Tue Jun 25
        type: string
      now_rel_time:
        example: 15 hours from now
        type: string
      now_time:
        example: 5 hours ago
        type: string
      rfc1123:
        example: Tue, 25 Jun 2019 03:00:53 AKDT
        type: string
      time:
        example: "03:00:53"
        type: string
      timezone:
        example: America/Anchorage
        type: string
      value:
        example: "2019-06-25T03:00:53.284-08:00"
        type: string
      value_utc:
        example: "2019-06-25T11:00:53.284Z"
        type: string
    type: object
host: '{{.Host}}'
info:
  contact:
    email: support@geeksinthewoods.com
    name: API Support
    url: http://example.com/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms
  title: SaaS Example API
  version: '{{.Version}}'
paths:
...
  /project:
    get:
      consumes:
      - application/json
      description: Find returns the existing projects in the system.
      parameters:
      - description: 'Filter string, example: name = ''Moon Launch'''
        in: query
        name: where
        type: string
      - description: 'Order columns separated by comma, example: created_at desc'
        in: query
        name: order
        type: string
      - description: 'Limit, example: 10'
        in: query
        name: limit
        type: integer
      - description: 'Offset, example: 20'
        in: query
        name: offset
        type: integer
      - description: 'Included Archived, example: false'
        in: query
        name: included-archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.ProjectResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: List projects
      tags:
      - project
  /projects:
    patch:
      consumes:
      - application/json
      description: Update updates the specified project in the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/project.ProjectUpdateRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Update project by ID
      tags:
      - project
    post:
      consumes:
      - application/json
      description: Create inserts a new project into the system.
      parameters:
      - description: Project details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/project.ProjectCreateRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Create new project.
      tags:
      - project
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete removes the specified project from the system.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Delete project by ID
      tags:
      - project
    get:
      consumes:
      - application/json
      description: Read returns the specified project from the system.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Get project by ID.
      tags:
      - project
  /projects/archive:
    patch:
      consumes:
      - application/json
      description: Archive soft-deletes the specified project from the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/project.ProjectArchiveRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Archive project by ID
      tags:
      - project
...
securityDefinitions:
  BasicAuth:
    type: basic
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants administrative privileges with role of admin.'
      user: ' Grants basic privileges with role of user.'
    tokenUrl: /v1/oauth/token
    type: oauth2
swagger: "2.0"